- Created `pyproject.toml` for modern Python packaging and dependency management. - Added `setup.py` for backward compatibility with the new packaging structure. - Refactored Snakemake workflow rules to utilize new script entry points instead of direct script paths.
