+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+
+# This setup.py file is kept for backwards compatibility
+# The actual configuration is in pyproject.toml
+
+setup(
+    name="AlleleFlux",
+    packages=["alleleflux"],
+)
