+++ smk_workflow/step1.smk
@@ -47,7 +47,6 @@ rule profile:
         time=config["time"]["profile"],
     params:
         outDir=os.path.join(OUTDIR, "profiles"),
-        scriptPath=config["scripts"]["profile"],
     benchmark:
         os.path.join(
             OUTDIR,
@@ -56,7 +55,7 @@ rule profile:
         )
     shell:
         """
-        python {params.scriptPath} \
+        alleleflux-profile \
         --bam_path {input.bam} --fasta_path {input.fasta} --prodigal_fasta {input.prodigal} \
         --cpus {threads} --output_dir {params.outDir}
         """
@@ -65,7 +64,7 @@ rule profile:
 rule generate_metadata:
     input:
         metadata=config["metadata_file"],
-        # This enures that generate_metadata is run after all samples are profiled
+        # This ensures that generate_metadata is run after all samples are profiled
         sampleDirs=expand(
             os.path.join(OUTDIR, "profiles", "{sample}.sorted"), sample=samples
         ),
@@ -77,7 +76,6 @@ rule generate_metadata:
         ),
     params:
         rootDir=os.path.join(OUTDIR, "profiles"),
-        scriptPath=config["scripts"]["generate_mag_metadata"],
         data_type=DATA_TYPE,
         group_args=lambda wildcards: f"--groups {wildcards.groups.replace('_', ' ')}",
         timepoint_args=lambda wildcards: f"--timepoints {wildcards.timepoints.replace('_', ' ')}",
@@ -85,7 +83,7 @@ rule generate_metadata:
         time=config["time"]["general"],
     shell:
         """
-        python {params.scriptPath} \
+        alleleflux-metadata \
             --rootDir {params.rootDir} \
             --metadata {input.metadata} \
             --outDir {output.outDir} \
@@ -105,14 +103,13 @@ rule qc:
         outDir=directory(os.path.join(OUTDIR, "QC", "QC_{timepoints}-{groups}")),
     params:
         breadth_threshold=config["breadth_threshold"],
-        scriptPath=config["scripts"]["quality_control"],
         data_type=DATA_TYPE,
     threads: config["cpus"]["quality_control"]
     resources:
         time=config["time"]["general"],
     shell:
         """
-        python {params.scriptPath} \
+        alleleflux-qc \
             --rootDir {input.metadata_dir} \
             --fasta {input.fasta} \
             --breadth_threshold {params.breadth_threshold} \
@@ -129,13 +126,12 @@ rule eligibility_table:
         out_fPath=os.path.join(OUTDIR, "eligibility_table_{timepoints}-{groups}.tsv"),
     params:
         min_sample_num=config["min_sample_num"],
-        script=config["scripts"]["eligibility_table"],
         data_type=DATA_TYPE,
     resources:
         time=config["time"]["general"],
     shell:
         """
-        python {params.script} \
+        alleleflux-eligibility \
             --qc_dir {input.qc_dir} \
             --min_sample_num {params.min_sample_num} \
             --output_file {output.out_fPath} \
